requirements.txt
pillow~=8.1.2
matplotlib~=3.3.4
numpy~=1.19.2